refactor(workflow-executor): harden bearer-claims validation after review [PRD-508]
- Narrow hasRunAccess(runId, user) to { id: number } (only the id is used), removing
  the unsound `as StepUser` cast on the validated bearer user — both routes now cast
  to BearerClaims consistently.
- Log invalid bearer claims (logger.warn, issue paths/codes only — never the payload)
  before the 401: a signed-but-malformed token is rare and high-signal, not churn.
- Tighten BearerClaimsSchema id to z.number().int() (rejects NaN/Infinity/floats),
  preserving the previous Number.isFinite invariant.
- Tests: GET route 401 on invalid claims (access check not reached), warn assertion,
  non-integer id rejection.

+++ packages/workflow-executor/src/http/bearer-claims.ts
@@ -3,6 +3,8 @@ import { z } from 'zod';
 // Claims we require from the decoded bearer JWT payload (ctx.state.user). Non-strict on purpose:
 // jsonwebtoken adds standard claims (iat/exp) and Forest may send extra ones — strict would reject
 // every real token. Only `id` is consumed downstream (handleTrigger + hasRunAccess → ?userId=).
-export const BearerClaimsSchema = z.object({ id: z.number() });
+// `.int()` rejects NaN/Infinity/floats (a user id is an integer) — preserves the invariant the
+// previous Number.isFinite guard enforced.
+export const BearerClaimsSchema = z.object({ id: z.number().int() });
 
 export type BearerClaims = z.infer<typeof BearerClaimsSchema>;
